refactor(backend): move domain pagination to 'shared' package, instead of 'model'
