+++ backend/core/src/main/java/com/wilzwert/myjobs/core/domain/model/user/ports/driving/GetUserJobsUseCase.java
@@ -2,7 +2,7 @@ package com.wilzwert.myjobs.core.domain.model.user.ports.driving;
 
 
 import com.wilzwert.myjobs.core.domain.model.job.EnrichedJob;
-import com.wilzwert.myjobs.core.domain.model.pagination.DomainPage;
+import com.wilzwert.myjobs.core.domain.shared.pagination.DomainPage;
 import com.wilzwert.myjobs.core.domain.model.job.JobStatus;
 import com.wilzwert.myjobs.core.domain.model.user.UserId;
