+++ backend/core/src/main/java/com/wilzwert/myjobs/core/domain/shared/pagination/DomainPage.java
@@ -1,4 +1,4 @@
-package com.wilzwert.myjobs.core.domain.model.pagination;
+package com.wilzwert.myjobs.core.domain.shared.pagination;
 
 
 import java.util.Collections;
