+++ backend/core/src/test/java/com/wilzwert/myjobs/core/domain/model/pagination/DomainPageTest.java
@@ -1,5 +1,6 @@
 package com.wilzwert.myjobs.core.domain.model.pagination;
 
+import com.wilzwert.myjobs.core.domain.shared.pagination.DomainPage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
