+++ backend/infrastructure/src/main/java/com/wilzwert/myjobs/infrastructure/mapper/EntityMapper.java
@@ -1,6 +1,6 @@
 package com.wilzwert.myjobs.infrastructure.mapper;
 
-import com.wilzwert.myjobs.core.domain.model.pagination.DomainPage;
+import com.wilzwert.myjobs.core.domain.shared.pagination.DomainPage;
 import com.wilzwert.myjobs.infrastructure.api.rest.dto.RestPage;
 import org.springframework.data.domain.Page;
